+++ Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/event/processor/IdentityDeleteProcessor.java
@@ -8,11 +8,11 @@ import org.springframework.util.Assert;
 import eu.bcvsolutions.idm.acc.dto.filter.AccIdentityAccountFilter;
 import eu.bcvsolutions.idm.acc.service.api.AccIdentityAccountService;
 import eu.bcvsolutions.idm.core.api.dto.IdmIdentityDto;
-import eu.bcvsolutions.idm.core.api.event.AbstractEntityEventProcessor;
 import eu.bcvsolutions.idm.core.api.event.CoreEvent;
 import eu.bcvsolutions.idm.core.api.event.DefaultEventResult;
 import eu.bcvsolutions.idm.core.api.event.EntityEvent;
 import eu.bcvsolutions.idm.core.api.event.EventResult;
+import eu.bcvsolutions.idm.core.api.event.processor.AbstractIdentityProcessor;
 import eu.bcvsolutions.idm.core.model.event.IdentityEvent.IdentityEventType;
 
 /**
@@ -23,7 +23,7 @@ import eu.bcvsolutions.idm.core.model.event.IdentityEvent.IdentityEventType;
  */
 @Component("accIdentityDeleteProcessor")
 @Description("Ensures referential integrity. Cannot be disabled.")
-public class IdentityDeleteProcessor extends AbstractEntityEventProcessor<IdmIdentityDto> {
+public class IdentityDeleteProcessor extends AbstractIdentityProcessor {
 	
 	public static final String PROCESSOR_NAME = "identity-delete-processor";
 	private final AccIdentityAccountService identityAccountService;