+++ Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/event/processor/RoleDeleteProcessor.java
@@ -10,11 +10,11 @@ import eu.bcvsolutions.idm.acc.dto.filter.SysRoleSystemFilter;
 import eu.bcvsolutions.idm.acc.service.api.AccRoleAccountService;
 import eu.bcvsolutions.idm.acc.service.api.SysRoleSystemService;
 import eu.bcvsolutions.idm.core.api.dto.IdmRoleDto;
-import eu.bcvsolutions.idm.core.api.event.AbstractEntityEventProcessor;
 import eu.bcvsolutions.idm.core.api.event.CoreEvent;
 import eu.bcvsolutions.idm.core.api.event.DefaultEventResult;
 import eu.bcvsolutions.idm.core.api.event.EntityEvent;
 import eu.bcvsolutions.idm.core.api.event.EventResult;
+import eu.bcvsolutions.idm.core.api.event.processor.AbstractRoleProcessor;
 import eu.bcvsolutions.idm.core.model.event.RoleEvent.RoleEventType;
 
 /**
@@ -25,7 +25,7 @@ import eu.bcvsolutions.idm.core.model.event.RoleEvent.RoleEventType;
  */
 @Component("accRoleDeleteProcessor")
 @Description("Ensures referential integrity. Cannot be disabled.")
-public class RoleDeleteProcessor extends AbstractEntityEventProcessor<IdmRoleDto> {
+public class RoleDeleteProcessor extends AbstractRoleProcessor {
 	
 	public static final String PROCESSOR_NAME = "role-delete-processor";
 	private final SysRoleSystemService roleSystemService;