+++ Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/event/processor/contract/AbstractIdentityContractProvisioningProcessor.java
@@ -5,11 +5,10 @@ import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
-import eu.bcvsolutions.idm.core.api.dto.IdmIdentityContractDto;
 import eu.bcvsolutions.idm.core.api.dto.IdmIdentityDto;
 import eu.bcvsolutions.idm.core.api.dto.filter.IdmIdentityFilter;
-import eu.bcvsolutions.idm.core.api.event.AbstractEntityEventProcessor;
 import eu.bcvsolutions.idm.core.api.event.CoreEvent.CoreEventType;
+import eu.bcvsolutions.idm.core.api.event.processor.AbstractIdentityContractProcessor;
 import eu.bcvsolutions.idm.core.api.service.IdmIdentityService;
 
 /**
@@ -17,7 +16,7 @@ import eu.bcvsolutions.idm.core.api.service.IdmIdentityService;
  * 
  * @author Radek Tomiška
  */
-public abstract class AbstractIdentityContractProvisioningProcessor extends AbstractEntityEventProcessor<IdmIdentityContractDto> {
+public abstract class AbstractIdentityContractProvisioningProcessor extends AbstractIdentityContractProcessor {
 
 	protected static final String PROPERTY_INCLUDE_SUBORDINATES = "includeSubordinates";
 	protected static final String PROPERTY_PREVIOUS_SUBORDINATES = "idm:previous-subordinates"; // contains Set<UUID>